requirements.txt
Django>=5.1.6
djangorestframework
python-dotenv
psycopg2-binary
gunicorn
drf_yasg
djangorestframework-simplejwt
django-cors-headers
torch
numpy
scikit-learn
requests
APScheduler
pandas
xgboost
matplotlib
seaborn
joblib